Make tests run in Node
This is a follow-up to
https://github.com/slevithan/xregexp/pull/155#issuecomment-279608958
that lets the tests run in Node, without opening a browser. It's set up
so that opening tests/index.html still works, for anyone who prefers
that workflow.

See the jasmine docs for more info on using it with Node: https://jasmine.github.io/2.5/node.html